refactor(core): Introduce interface for StageResolver
