+++ orca-core/src/main/java/com/netflix/spinnaker/orca/config/OrcaConfiguration.java
@@ -24,6 +24,7 @@ import com.netflix.spinnaker.config.PluginsAutoConfiguration;
 import com.netflix.spinnaker.kork.core.RetrySupport;
 import com.netflix.spinnaker.kork.dynamicconfig.DynamicConfigService;
 import com.netflix.spinnaker.kork.expressions.ExpressionFunctionProvider;
+import com.netflix.spinnaker.orca.DefaultStageResolver;
 import com.netflix.spinnaker.orca.StageResolver;
 import com.netflix.spinnaker.orca.TaskResolver;
 import com.netflix.spinnaker.orca.api.pipeline.Task;
@@ -158,7 +159,8 @@ public class OrcaConfiguration {
 
   @Bean
   @ConditionalOnMissingBean(StageDefinitionBuilderFactory.class)
-  public StageDefinitionBuilderFactory stageDefinitionBuilderFactory(StageResolver stageResolver) {
+  public StageDefinitionBuilderFactory stageDefinitionBuilderFactory(
+      DefaultStageResolver stageResolver) {
     return new DefaultStageDefinitionBuilderFactory(stageResolver);
   }
 
@@ -211,11 +213,12 @@ public class OrcaConfiguration {
   }
 
   @Bean
-  public StageResolver stageResolver(
+  @ConditionalOnMissingBean(StageResolver.class)
+  public DefaultStageResolver defaultStageResolver(
       Collection<StageDefinitionBuilder> stageDefinitionBuilders,
       Optional<List<SimpleStage>> simpleStages) {
     List<SimpleStage> stages = simpleStages.orElseGet(ArrayList::new);
-    return new StageResolver(stageDefinitionBuilders, stages);
+    return new DefaultStageResolver(stageDefinitionBuilders, stages);
   }
 
   @Bean(name = EVENT_LISTENER_FACTORY_BEAN_NAME)
